relocated login logout buttons
