+++ client/src/components/Navbar.js
@@ -20,8 +20,8 @@ import { ThemeProvider } from '@emotion/react';
 import DefaultTheme from './DefaultTheme';
 import DarkTheme from './DarkTheme';
 import { useThemeContext } from './ThemeContext';
-import LoginButton from '../LoginButton';
-import LogoutButton from '../LogoutButton';
+import LoginButton from './LoginButton';
+import LogoutButton from './LogoutButton';
 import { useAuth0 } from "@auth0/auth0-react";
 import { useUserContext } from './UserContext';
 import ReservationsMenu from './ReservationsDropDown';
